+++ pyembed/src/test/interpreter_config.rs
@@ -35,6 +35,7 @@ rusty_fork_test! {
         let mut config = OxidizedPythonInterpreterConfig::default();
         // Otherwise the Rust arguments are interpreted as Python arguments.
         config.interpreter_config.parse_argv = Some(false);
+        config.set_missing_path_configuration = false;
         let mut interp = MainPythonInterpreter::new(config).unwrap();
 
         let py = interp.acquire_gil().unwrap();
@@ -64,7 +65,7 @@ rusty_fork_test! {
         // detected by python3_sys. Without this, sys.path and other paths reference
         // directories next to the Rust test executable, and there is no Python stdlib
         // there.
-        config.isolated_auto_set_path_configuration = false;
+        config.set_missing_path_configuration = false;
 
         let mut interp = MainPythonInterpreter::new(config).unwrap();
 
@@ -113,6 +114,7 @@ rusty_fork_test! {
         let mut config = OxidizedPythonInterpreterConfig::default();
         // Otherwise the Rust arguments are interpreted as Python arguments.
         config.interpreter_config.parse_argv = Some(false);
+        config.set_missing_path_configuration = false;
 
         let mut interp = MainPythonInterpreter::new(config).unwrap();
 
@@ -134,6 +136,7 @@ rusty_fork_test! {
         let mut config = OxidizedPythonInterpreterConfig::default();
         // Otherwise the arguments are interpreted as Python arguments.
         config.interpreter_config.parse_argv = Some(false);
+        config.set_missing_path_configuration = false;
 
         // .argv expands to current process args by default. But setting
         // .interpreter_config.argv overrides this behavior.
@@ -163,6 +166,7 @@ rusty_fork_test! {
         let mut config = OxidizedPythonInterpreterConfig::default();
         // Otherwise the arguments are interpreted as Python arguments.
         config.interpreter_config.parse_argv = Some(false);
+        config.set_missing_path_configuration = false;
         config.argv = Some(vec![
             OsString::from("prog"),
             OsString::from("foo"),
@@ -188,6 +192,7 @@ rusty_fork_test! {
 
         // Otherwise the Rust arguments are interpreted as Python arguments.
         config.interpreter_config.parse_argv = Some(false);
+        config.set_missing_path_configuration = false;
         config.argv = Some(vec![get_unicode_argument()]);
         config.argvb = true;
 
@@ -220,6 +225,7 @@ rusty_fork_test! {
 
         // Otherwise the Rust arguments are interpreted as Python arguments.
         config.interpreter_config.parse_argv = Some(false);
+        config.set_missing_path_configuration = false;
         config.argv = Some(vec![get_unicode_argument()]);
 
         let mut interp = MainPythonInterpreter::new(config).unwrap();
@@ -258,7 +264,8 @@ rusty_fork_test! {
         config.interpreter_config.profile = PythonInterpreterProfile::Isolated;
         // Otherwise the Rust arguments are interpreted as Python arguments.
         config.interpreter_config.parse_argv = Some(false);
-        config.isolated_auto_set_path_configuration = false;
+        config.set_missing_path_configuration = false;
+        config.set_missing_path_configuration = false;
         config.argv = Some(vec![get_unicode_argument()]);
 
         let mut interp = MainPythonInterpreter::new(config).unwrap();
@@ -321,7 +328,7 @@ rusty_fork_test! {
         config.interpreter_config.configure_locale = Some(true);
         // Otherwise the Rust arguments are interpreted as Python arguments.
         config.interpreter_config.parse_argv = Some(false);
-        config.isolated_auto_set_path_configuration = false;
+        config.set_missing_path_configuration = false;
         config.argv = Some(vec![get_unicode_argument()]);
 
         let mut interp = MainPythonInterpreter::new(config).unwrap();