+++ pyoxidizer/src/py_packaging/config.rs
@@ -186,7 +186,7 @@ impl EmbeddedPythonConfig {
             x_options: {},\n        \
             }},\n    \
             raw_allocator: Some({}),\n    \
-            isolated_auto_set_path_configuration: true,\n    \
+            set_missing_path_configuration: true,\n    \
             oxidized_importer: {},\n    \
             filesystem_importer: {},\n    \
             packed_resources: {},\n    \